requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
black==21.5b1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
flake8==3.9.2
Flask==2.0.1
idna==2.10
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.8.1
pycodestyle==2.7.0
pyflakes==2.3.1
PySocks==1.7.1
python-telegram-bot==13.6
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.5
waitress==2.1.1
Werkzeug==2.0.1
